Move wallbox coordinator to its own file
